\name{dugen}
\alias{dugen}
\title{Uniform Numbers}
\usage{
  dugen(st,en,num)
}
\description{
Using "rgenerator","dugen" is a function, taking the two integer numbers as input and
generates random numbers uniformly distributed between two integer inputs.

related functions :
	runif

}
\examples{
dugen(10,20,100)
}
