\name{gegen}
\alias{gegen}
\title{Geometric Distribution}
\usage{
  gengen(p)
}
\description{
The number of failures in Bernoulli trials, between two wins, follows the geometric distributions.
Using brgen,gegen is a function, taking a parameters and generates a geometric random variable.

related functions :
	rgeom

}
\examples{
gegen(0.27)
}
