\name{visualizebi}
\alias{visualizebi}
\title{Binomial Numbers histogram Visualization }
\usage{
  visualizebi(p,num , brick , color)
}
\description{
this function generates 1000 Binomial random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizebi(0.27,10,10,"red")
}

