\name{visualizebiPlot}
\alias{visualizebiPlot}
\title{Binomial Numbers Visualization}
\usage{
  visualizebiPlot(p,num)
}
\description{
this function generates 1000 binomial random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizebiPlot(0.27,10)
}

