\name{visualizebr}
\alias{visualizebr}
\title{Bernoulli Numbers histogram Visualization 
\usage{
  visualizebr(p , brick , color)
}
\description{
this function generates 1000 Bernoulli random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizebr(0.27,10,"blue")
}
