\name{visualizebrPlot}
\alias{visualizebrPlot}
\title{Bernoulli Numbers ggplot Visualization }
\usage{
  visualizebrPlot(p)
}
\description{
this function generates 1000 bernoulli random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizebrPlot(0.27)
}
