\name{visualizeDu}
\alias{visualizeDu}
\title{Uniform Numbers histogram Visualization}
\usage{
  visualizeDu(st,en , brick , color)
}
\description{
this function generates 10000 dugen random numbers and then shows the histogram plot for the generated random numbers.
}
\examples{
visualizeDu(0,10,10,"blue")
}
