\name{visualizeDuPlot}
\alias{visualizeDuPlot}
\title{Uniform Numbers ggplot Visualization}
\usage{
  visualizeDuPlot(st , en)
}
\description{
this function generates 10000 dugen random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizeDuPlot(0,10)
}
