\name{visualizeEXP}
\alias{visualizeEXP}
\title{Exponential Numbers histogram Visualization}
\usage{
  visualizeEXP(lambda , brick , color)
}
\description{
this function generates 1000 Exponential random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizeEXP(10,10,"red")
}

