\name{visualizeEXPPlot}
\alias{visualizeEXPPlot}
\title{Exponential Numbers ggplot Visualization}
\usage{
  visualizeEXPPlot(lambda)
}
\description{
this function generates 1000 exponential random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizeEXPPlot(10)
}

