\name{visualizeGAMMA}
\alias{visualizeGAMMA}
\title{Gamma Numbers histogram Visualization}
\usage{
  visualizeGAMMA(lambda, k , brick , color)
}
\description{
this function generates 1000 Gamma random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizeGAMMA(10,3,10,"black")
}