\name{visualizeGAMMAPlot}
\alias{visualizeGAMMAPlot}
\title{Gamma Numbers ggplot Visualization}
\usage{
  visualizeGAMMAPlot(lambda, k )
}
\description{
this function generates 1000 gamma random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizeGAMMAPlot(10,3)
}