\name{visualizegeo}
\alias{visualizegeo}
\title{Geometric Numbers histogram Visualization}
usage{
  visualizegeo(p , brick , color)
}
\description{
this function generates 1000 Geometric random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizegeo(0.27,10,"white")
}
