\name{visualizegeoPlot}
\alias{visualizegeoPlot}
\title{Geometric Numbers ggplot Visualization}
\usage{
  visualizegeoPlot(p )
}
\description{
this function generates 1000 geometeric random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizegeoPlot(0.27)
}
