\name{visualizeNORMAL}
\alias{visualizeNORMAL}
\title{Normal Numbers histogram Visualization}
\usage{
  visualizeNORMAL(u, s , brick , color)
}
\description{
this function generates 1000 Normal random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizeNORMAL(20,20,10,"blue")
}
