\name{visualizeNORMALPlot}
\alias{visualizeNORMALPlot}
\title{Normal Numbers ggplot Visualization}
\usage{
  visualizeNORMALPlot(u, s )
}
\description{
this function generates 1000 normal random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizeNORMALPlot(20,20)
}
