\name{visualizePOISSON}
\alias{visualizePOISSON}
\title{Poisson Numbers histogram Visualization}
\usage{
  visualizePOISSON(lambda, t , brick , color)
}
\description{
this function generates 1000 Poisson random numbers and then shows the histogram plot for the generated random numbers using ggplot.
}
\examples{
visualizePOISSON(10,20,10,"red")
}