\name{visualizePOISSONPlot}
\alias{visualizePOISSONPlot}
\title{Poisson Numbers ggplot Visualization}
\usage{
  visualizePOISSONPlot(lambda, t)
}
\description{
this function generates 1000 poisson random numbers and then shows the plot for the pdf of generated random numbers using ggplot.
}
\examples{
visualizePOISSONPlot(10,20)
}